+++ src/test/java/org/bot/nullbot/NullBotApplicationTests.java
@@ -91,7 +91,7 @@ class NullBotApplicationTests
        String commandType = "help";
        List<String> commandParameters =  new ArrayList<>();
        commandProcessor.processTest(new CommandEvent<>(
-               commandType, commandParameters, new GroupMessageEvent(),
+               new GroupMessageEvent(), commandType, commandParameters,
                false, true)
        );